requirements.txt
python-telegram-bot==13.7
py2app==0.24
PyYAML==5.4.1